disable all http instrumentation to fix hanging tests
The issue was that the new setTimeout() handler added in the latest
@nui/openwhisk-newrelic 5.0.0 when a response is not read is very often
triggered by the tests here (api.js, http multipart, shell script).
It seems they don't seem to always read the response properly, and
use of nock might also be a factor.

As a simple fix, I added env vars to completely disable the
instrumentation. It isn't really needed at all here.